Use low-level graph instead
